rework server, and simLoad function so all in #1
