+++ server/h_ajax.go
@@ -89,51 +89,3 @@ func H_AjaxTrackItems(w http.ResponseWriter, r *http.Request) {
 
 	SendJson(w, data)
 }
-
-/*
-H_SvgImageTest()  experimental to show svg
-*/
-func H_SvgImageTest(w http.ResponseWriter, r *http.Request) {
-
-	w.Header().Set("Content-Type", "image/svg+xml")
-	s := svg.New(w)
-	s.Start(1000, 800)
-
-
-	// Add circle for trackitem.. We Dont have __type__ ;-(
-	for _, ti := range sim.TrackItems {
-
-
-		switch ti.Type(){
-		case "LineItem":
-			s.Circle(ti.Origin().Xi(), ti.Origin().Yi(), 2, "fill:none;stroke:black")
-			s.Line(ti.Origin().Xi(),ti.Origin().Yi(), 20, 20,  "fill:none;stroke:green")
-
-			//s.Line(ti.Origin().Xi(),ti.Origin().Yi(), ti.End().Xi(), 20,  "fill:none;stroke:green")
-			// WTF - how to get END!!!
-			//m := structs.Map(ti)
-			//xe, _ := m["Xf"].(int)
-			//ye, _ := m["Yf"].(int)
-			//fmt.Println("==",xe,ye)
-			//s.Line(ti.Origin().Xi(),ti.Origin().Yi(), xe, ye,  "fill:none;stroke:green")
-
-		case "SignalItem":
-			s.Circle(ti.Origin().Xi(), ti.Origin().Yi(), 5, "fill:none;stroke:red")
-
-		default:
-			s.Circle(ti.Origin().Xi(), ti.Origin().Yi(), 2, "fill:none;stroke:grey")
-
-		}
-
-
-	}
-	s.End()
-}
-
-func H_SvgTest(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	tpl := getTemplate("templates/svg.html")
-	var data interface{}
-	tpl.Execute(w, data)
-	log.Println("HTML: /svg")
-}
